requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.9.0
apify-client==1.6.1
apify_shared==1.1.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-substrate-interface==1.5.1
async-timeout==5.0.1
asyncpraw==7.8.1
asyncprawcore==2.4.0
asyncstdlib==3.13.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.13.4
bittensor==9.9.0
bittensor-cli==9.8.4
bittensor-commit-reveal==0.4.0
bittensor-drand==1.0.0
bittensor-wallet==4.0.0
bleach==6.2.0
bt-decode==0.6.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
datasets==2.20.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.110.3
fastjsonschema==2.21.1
filelock==3.18.0
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2024.5.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-api-python-client==2.167.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.1
grpcio-tools==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy-opentensor==0.5.0
multidict==6.4.4
multiprocess==0.70.16
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.4.2
numpy==2.0.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
password-strength==0.0.3.post2
pexpect==4.9.0
platformdirs==4.3.8
playwright==1.52.0
playwright-stealth==1.0.6
plotille==5.0.0
plotly==6.1.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-bip39-bindings==0.1.11
pyarrow==17.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
python-statemachine==2.5.0
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==26.4.0
qtconsole==5.6.1
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.4
retry==0.9.2
rich==13.7.0
rpds-py==0.25.1
rsa==4.9.1
scalecodec==1.2.11
schedule==1.2.2
Send2Trash==1.8.3
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
wandb==0.18.7
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
youtube-transcript-api==1.2.0
langdetect
langid
aiohttp_socks
prometheus-fastapi-instrumentator==7.1.0
substrate-interface==1.7.11
